Consolidate upload/worker.ts into one sync file
